requirements.txt
pymongo == 3.11.1
dnspython==2.3.0
python-telegram-bot==20.3
pydantic==1.10.9